+++ packages/core/globals/index.ts
@@ -359,5 +359,5 @@ export function initGlobal() {
 	}
 }
 if (!global.NativeScriptHasInitGlobal) {
-    initGlobal();
+	initGlobal();
 }