+++ packages/core/ui/time-picker/index.d.ts
@@ -54,6 +54,16 @@ export class TimePicker extends View {
 	 * Gets or sets the minute interval.
 	 */
 	minuteInterval: number;
+
+	/**
+	 * Gets or set the UIDatePickerStyle of the date picker in iOS 13.4+. Defaults to 0.
+	 * Valid values are numbers:
+	 *  - 0: automatic (system picks the concrete style based on the current platform and date picker mode)
+	 *  - 1: wheels (the date picker displays as a wheel picker)
+	 *  - 2: compact (the date picker displays as a label that when tapped displays a calendar-style editor)
+	 *  - 3: inline  (the date pickers displays as an inline, editable field)
+	 */
+	iosPreferredDatePickerStyle: number;
 }
 
 export const hourProperty: Property<TimePicker, number>;
@@ -66,3 +76,5 @@ export const minMinuteProperty: Property<TimePicker, number>;
 
 export const timeProperty: Property<TimePicker, Date>;
 export const minuteIntervalProperty: Property<TimePicker, number>;
+
+export const iosPreferredDatePickerStyleProperty: Property<TimePicker, number>;